+++ CHANGELOG.md
@@ -14,6 +14,14 @@ Running changelog of releases since `2.0.0-rc.4`
   - Corrected env var typo on `OKTA_CLIENT_REQUEST_TIMEOUT` in README
   - Corrected code generation pass by referrence of a Factor instance in `EnrollFactor`, `ActivateFactor`, and `VerifyFactor`
   - Provider query parameter can be boolean or string
+  - Corrected out of order method and API parameters
+    - `AuthorizationServerResource`
+      - `DeleteAuthorizationServerPolicyRule(ctx context.Context, authServerId string, policyId string, ruleId string)`
+      - `GetAuthorizationServerPolicyRule(ctx context.Context, authServerId string, policyId string, ruleId string)`
+      - `UpdateAuthorizationServerPolicyRule(ctx context.Context, authServerId string, policyId string, ruleId string, body AuthorizationServerPolicyRule)`
+    - `AuthorizationServerPolicyRuleResource`
+      - `UpdateAuthorizationServerPolicyRule(ctx context.Context, authServerId string, policyId string, ruleId string, body AuthorizationServerPolicyRule)`
+      - `DeleteAuthorizationServerPolicyRule(ctx context.Context, authServerId string, policyId string, ruleId string)`
 
 ## v2.11.2